+++ src/impl/utils.hpp
@@ -13,6 +13,8 @@
 
 #include <climits>
 #include <limits>
+#include <list>
+#include <map>
 #include <random>
 #include <vector>
 
@@ -31,6 +33,12 @@ string base64_encode(const binary &data);
 
 // Return a random seed sequence
 std::seed_seq random_seed();
+
+// Parse an http packet into lines
+size_t parseHttpLines(const byte *buffer, size_t size, std::list<string> &lines);
+
+// Parse headers of a http message
+std::multimap<string, string> parseHttpHeaders(const std::list<string> &lines);
 
 template <typename Generator, typename Result = typename Generator::result_type>
 struct random_engine_wrapper {
