+++ src/impl/wshandshake.hpp
@@ -52,8 +52,6 @@ public:
 private:
 	static string generateKey();
 	static string computeAcceptKey(const string &key);
-	static size_t parseHttpLines(const byte *buffer, size_t size, std::list<string> &lines);
-	static std::multimap<string, string> parseHttpHeaders(const std::list<string> &lines);
 
 	string mHost;
 	string mPath;
